Fix issues reported by Phan

+++ application/inc/Entity/AbstractEntity.php
@@ -114,6 +114,10 @@ abstract class AbstractEntity implements InterfaceEntity
      */
     public function delete(): bool
     {
+        if ($this->id === null) {
+            return true;
+        }
+
         db()->query("DELETE FROM `" . static::TABLE_NAME . "` WHERE `id` = " . $this->id);
         ORM::forget(static::class, $this->id);
